refactor: introduce pkg/kiosk/config package with type aliases for migration
